refactor setLife to be pure function module, not Redux reducer
